Fix geolocation API returning wrong coordinates
When passing actual latitude/longitude pairs to the
Geolocation API it seems to not return those coordinates
but some other coordinates close by. Thus this commit
implements a check for directly providing coordinates, and
if that is the case, does not use the geolocation api.

A serious downside of this approach is that we're relying
on a specific geolocation format, i.e. "latitude,longitude".
An alternative could be to use two "data-" attributes for
latitude and longitude?